refactor:  merge slice directories into a single folder structure, fix restful name convention
