+++ app/[...rest]/page.tsx
@@ -2,10 +2,10 @@ import RestForm from '@/components/RestClient';
 import ResponseSection from '@/components/Response';
 import { ResizableGroup, ResizableHandle, ResizablePanel } from '@/components/Resizable';
 
-export default function RESTfull() {
+export default function RESTful() {
   return (
     <main className="flex h-[calc(100vh-6rem)]">
-      <ResizableGroup autoSaveId="restfull" direction="vertical">
+      <ResizableGroup autoSaveId="restful" direction="vertical">
         <ResizablePanel defaultSizePercentage={50} minSizePercentage={15}>
           <RestForm />
         </ResizablePanel>
